requirements.txt
langgraph
langchain
langchain-openai
psycopg2-binary
pgvector
redis
fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic-settings
pyyaml
openai   # using OpenRouter-compatible client
sqlalchemy[asyncio]
asyncpg
websockets>=12.0
fastapi-sso~=0.10.0
jsonschema>=4.20.0
requests>=2.31.0

# Observability - Metrics & Logging
structlog>=24.1.0
python-json-logger>=2.0.7
prometheus-fastapi-instrumentator>=7.0.0

# Observability - Distributed Tracing
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-instrumentation-requests>=0.41b0
opentelemetry-instrumentation-sqlalchemy>=0.41b0
opentelemetry-instrumentation-psycopg2>=0.41b0
opentelemetry-exporter-otlp-proto-grpc>=1.20.0
opentelemetry-exporter-otlp-proto-http>=1.20.0
tiktoken>=0.5.0  # Token counting for chunking
pymupdf>=1.23.0  # PDF text extraction (better than PyPDF2)

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.24.0
faker>=19.0.0
aiosqlite>=0.19.0
